Implement async support (#9)
Currently, the calculation script runs synchronously, meaning it will block the editor UI until it completes.

This PR adds an async support. The script will run in the background and the results will be populated when it'll finish. This features support **both** NeoVim and Vim 8.

When using an unsupported version, or when setting `g:import_cost_disable_async` to `1`, the synchronous execution will be used.

In addition, `g:import_cost_always_open_split` is set to `1` by default, since the command can be used in an autocommand (e.g. `autocmd InsertLeave * ImportCost`), and the expected behaviour is to always update the results buffer even if there is a single import.